Reverted part of 7f1b522 due to a bug leading to infinite cost

+++ src/routes-lib/bezier/bezier.cpp
@@ -4,15 +4,15 @@
 
 #include "bezier.h"
 
-std::map<int, std::vector<long>> Bezier::binomial_coeffs;
+std::map<int, std::vector<int>> Bezier::binomial_coeffs;
 
-const std::vector<long>& Bezier::getBinomialCoefficients(int degree) {
+const std::vector<int>& Bezier::getBinomialCoefficients(int degree) {
 
     // These will be reused several times so there is no point in recalculating them
     if (!binomial_coeffs.count(degree)) {
         
         // We use degree + 1 because a bezier curve has degree + 1 terms
-        binomial_coeffs[degree] = std::vector<long>(degree + 1);
+        binomial_coeffs[degree] = std::vector<int>(degree + 1);
         
         // Fill it
         for (int i = 0; i < degree + 1; i++)
@@ -28,7 +28,7 @@ glm::vec3 Bezier::evaluateBezierCurve(const std::vector<glm::vec3>& points, floa
 
     // Get the degree of the curve and the binomials that correspond to it 
     int degree = points.size() - 1;
-    const std::vector<long>& binoms = getBinomialCoefficients(degree);
+    const std::vector<int>& binoms = getBinomialCoefficients(degree);
 
     float one_minus_s = 1.0 - s;
     glm::vec3 point = glm::vec3(0.0);
@@ -49,7 +49,7 @@ std::vector<glm::vec3> Bezier::evaluateEntireBezierCurve(const std::vector<glm::
 
     // Get the degree of the curve and the binomials that correspond to it 
     int degree = points.size() - 1;
-    const std::vector<long>& binoms = getBinomialCoefficients(degree);
+    const std::vector<int>& binoms = getBinomialCoefficients(degree);
 
     // Figure out how far along the curve each point is. We use num_desired - 1 as the divisor so that we make sure we evaluate at 1
     std::vector<glm::vec3> points_calc = std::vector<glm::vec3>(num_desired);
@@ -88,7 +88,7 @@ float Bezier::bezierLength(const std::vector<glm::vec3>& points) {
 
 
 
-long Bezier::calcBinomialCoefficient(int n, int i) {
+int Bezier::calcBinomialCoefficient(int n, int i) {
 
     // Special case to prevent divide by zero
     if (!i || n == i)
